% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridProcessing.R
\name{classify_responses}
\alias{classify_responses}
\title{Title}
\usage{
classify_responses(grid)
}
\arguments{
\item{grid}{a matrix.}
}
\value{
A list containing the grid cell counts for each of four categories:
  indifferent, ambivalent, positive, and negative.
}
\description{
Title
}
