% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridProcessing.R
\name{create_grid_score}
\alias{create_grid_score}
\title{Title}
\usage{
create_grid_score(x, gridinfo, b = -0.5, reverse_code = NULL)
}
\arguments{
\item{x}{}

\item{gridinfo}{}

\item{reverse_code}{}
}
\value{

}
\description{
Title
}
