% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridProcessing.R
\name{delete_empty_grid}
\alias{delete_empty_grid}
\title{Title}
\usage{
delete_empty_grid(resp_list, remove = TRUE)
}
\arguments{
\item{resp_list}{list of matrices.}

\item{remove}{logical. If \code{TRUE} empty matrices are removed from the list.}
}
\value{

}
\description{
Title
}
