% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridProcessing.R
\name{fix_limesurvey_likert}
\alias{fix_limesurvey_likert}
\title{Title}
\usage{
fix_limesurvey_likert(
  dat,
  lsvals = c("D4", "D3", "D2", "D1", "N0", "A1", "A2", "A3", "A4"),
  livals = 1:9
)
}
\arguments{
\item{dat}{data frame of LimeSurvey formatted Likert-type data.}

\item{lsvals}{character vector of LimeSurvey's notation for Likert-type scale points.}

\item{livals}{integer vector of Likert-type scale points.}
}
\value{
A data frame of numeric Likert-type values for each response.
}
\description{
Title
}
