% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrinomialTest.R
\name{gen.probs.obj}
\alias{gen.probs.obj}
\title{Title}
\usage{
gen.probs.obj(
  n,
  alpha = 0.05,
  include.one = TRUE,
  find.RR = TRUE,
  digits = NULL,
  P0s = NULL
)
}
\arguments{
\item{n}{}

\item{alpha}{}

\item{include.one}{}

\item{find.RR}{}

\item{digits}{}

\item{P0s}{}
}
\value{

}
\description{
Title
}
