% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrinomialTest.R
\name{generate_trinomial_info}
\alias{generate_trinomial_info}
\title{Title}
\usage{
generate_trinomial_info(
  n,
  alpha = 0.05,
  include.one = TRUE,
  find.RR = TRUE,
  digits = NULL,
  P0s = NULL
)
}
\arguments{
\item{n}{}

\item{alpha}{}

\item{include.one}{}

\item{find.RR}{Logical; determines if the Rejection Region should be found.}

\item{digits}{}

\item{P0s}{}
}
\value{

}
\description{
Title
}
