% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridProcessing.R
\name{grid.cell.counts}
\alias{grid.cell.counts}
\title{Title}
\usage{
grid.cell.counts(x, gridinfo, type = "items", return.table = FALSE)
}
\arguments{
\item{x}{}

\item{gridinfo}{}

\item{type}{}

\item{return.table}{}
}
\value{

}
\description{
Title
}
