% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridProcessing.R
\name{grid.tri.summary}
\alias{grid.tri.summary}
\title{Title}
\usage{
grid.tri.summary(mat, rows = 5, cols = 5, offdiag = 0, return.table = TRUE)
}
\arguments{
\item{mat}{}

\item{rows}{}

\item{cols}{}

\item{offdiag}{}

\item{return.table}{}
}
\value{

}
\description{
Title
}
