% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridProcessing.R
\name{grid_item_info}
\alias{grid_item_info}
\title{Show Grid Item Information}
\usage{
grid_item_info(x, rows = 5, cols = 5)
}
\arguments{
\item{x}{data frame of raw LimeSurvey formatted data.}

\item{rows, cols}{the dimensions of the grid items.}
}
\value{
A list containing information on which raw data columns contain grid items, 
  as well as the item names and dimensions of the grid items.
}
\description{
Displays the LimeSurvey columns, item name, and dimensions of each grid item.
}
