% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridProcessing.R
\name{grid_summary_tri}
\alias{grid_summary_tri}
\title{Title}
\usage{
grid_summary_tri(
  mat,
  rows = 5,
  cols = 5,
  offdiag = 0,
  return_table = TRUE,
  limesurvey = TRUE
)
}
\arguments{
\item{mat}{a matrix.}

\item{rows, cols}{the dimensions of the matrix.}

\item{offdiag}{integer. the number of off-diagonals to treat as part of the diagonal.}

\item{return_table}{logical. If \code{TRUE} the function returns a table rather than a list.}

\item{limesurvey}{logical. If \code{FALSE} the function uses the sum of the diagonal from bottom-left to top-right.}
}
\value{
A list (or table) containing the grid cell counts for each of three categories:
  above the diagonal, on the diagonal, and below the diagonal.
}
\description{
Title
}
