% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridProcessing.R
\name{grid_to_uni}
\alias{grid_to_uni}
\title{Convert Grid Value to Likert-type Value}
\usage{
grid_to_uni(
  gc,
  rows = 5,
  cols = 5,
  b = (uni.min - uni.max)/16,
  uni.min = 1,
  uni.max = 9,
  rc = FALSE
)
}
\arguments{
\item{gc}{integer indicating a grid cell's position as though the grid were a
matrix filled by row with sequential integers.}

\item{rows, cols}{the dimensions of the grid item.}

\item{b}{a parameter between -1 and 0, default is a value satisfying Audrezet's (2016) six constraints.}

\item{uni.min, uni.max}{integers specifying the minimum and maximum value for the unidimensional (Likert-type) mapping. Default is a 9-point scale.}

\item{rc}{logical. If \code{TRUE} the function will correct for a reverse-coded grid item.}
}
\value{
A number derived from a grid cell's position
  and transformed to be equivalent to a value on a 9 point Likert-type scale.
}
\description{
Converts a grid cell's position to a score equivalent to a Likert-type scale point.
}
