% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridProcessing.R
\name{grid_trace}
\alias{grid_trace}
\title{Sum of the Diagonal}
\usage{
grid_trace(mat, col = NULL, limesurvey = TRUE)
}
\arguments{
\item{mat}{a square matrix.}

\item{col}{the number of columns of the matrix.}

\item{limesurvey}{logical. If \code{FALSE} the function calculates the sum of the diagonal from bottom-left to top-right.}
}
\value{
The sum of the matrix's diagonal, either from top-left to bottom-right 
  (the default) or from bottom-left to top-right.
}
\description{
Returns the sum of all cells on either a grid's main diagonal (trace) or its secondary diagonal.
}
