% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{griddata_all_text}
\alias{griddata_all_text}
\title{All Text Grid Data}
\format{
An object of class \code{data.frame} with 42 rows and 646 columns.
}
\usage{
griddata_all_text
}
\description{
All Text Grid Data
}
\author{
Douglas Whitaker
}
\keyword{datasets}
