% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{griddata_gridonly_code}
\alias{griddata_gridonly_code}
\title{Grid-only Grid Data}
\format{
An object of class \code{data.frame} with 42 rows and 551 columns.
}
\usage{
griddata_gridonly_code
}
\description{
Grid-only Grid Data
}
\author{
Douglas Whitaker
}
\keyword{datasets}
