% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphingFunctions.R
\name{make_heatmap}
\alias{make_heatmap}
\title{Make a Heatmap for a Grid}
\usage{
make_heatmap(
  grid,
  labels = c("agree_disagree", "satisfied_dissatisfied", "positive_negative",
    "accurate_inaccurate", "other"),
  title = "Heatmap for Grid Item",
  pos_labels = NULL,
  neg_labels = NULL,
  colours = c("purple", "green", "red", "custom"),
  custom_palette,
  show_counts = TRUE,
  breaks = NA,
  fontsize = 20,
  fontsize_names = 11
)
}
\arguments{
\item{grid}{a matrix.}

\item{labels}{a character string specifying which set of row names and col names
are to be used.}

\item{title}{the title of the plot.}

\item{pos_labels}{character. Custom column names.}

\item{neg_labels}{character. Custom row names.}

\item{colours}{a character string indicating what colour the heatmap will be.
Choosing \code{"custom"} allows for setting the \code{custom_palette} variable
to a palette other than \code{"purple"}, \code{"green"}, or \code{"red"}.}

\item{custom_palette}{a colour palette, the output of \code{colorRampPalette}.}

\item{show_counts}{logical. If \code{TRUE}, counts are shown for each cell.}

\item{breaks}{A sequence of numbers that covers the range of values in the grid.
If the value is \code{NA}, the breaks are calculated automatically.}

\item{fontsize}{base fontsize for the heatmap.}

\item{fontsize_names}{fontsize for row and column names.}
}
\value{
a heatmap showing counts in each grid cell.
}
\description{
The function makes a heatmap showing cell counts for a grid. By default, 
darker colours indicate larger counts, while lighter colours indicate smaller counts.
}
