% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridProcessing.R
\name{sum_grid_mat_list}
\alias{sum_grid_mat_list}
\title{Sum of Grids}
\usage{
sum_grid_mat_list(mat_list, items = NULL)
}
\arguments{
\item{mat_list}{list of matrices.}

\item{items}{indicates which matrices to sum.}
}
\value{
A matrix created by summing some or all of the inputted list of grids.
}
\description{
Sums across all (or a selection) of grids in a list.
}
