% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridProcessing.R
\name{sum_resp_mats}
\alias{sum_resp_mats}
\title{Title}
\usage{
sum_resp_mats(mat_list, items = NULL)
}
\arguments{
\item{mat_list}{list of matrices.}

\item{items}{indicates which matrices to sum.}
}
\value{
A matrix created by summing some or all of the inputted list of grids.
}
\description{
Title
}
