% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrinomialTest.R
\name{trinomial.test}
\alias{trinomial.test}
\title{Perform the trinomial hypothesis test}
\usage{
trinomial.test(
  col1,
  col2,
  alternative = c("two.sided", "greater", "less"),
  p_tie = NULL
)
}
\arguments{
\item{col1}{first data vector}

\item{col2}{second data vector}

\item{alternative}{the type of hypothesis test to be performed}

\item{p_tie}{theoretical proportion of ties; calculated from the data by default}
}
\value{
the p-value from the hypothesis test
}
\description{
Perform the trinomial hypothesis test
}
