% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridProcessing.R
\name{within_diag}
\alias{within_diag}
\title{Sum of the Diagonal and Off-diagonals}
\usage{
within_diag(mat, col = 5)
}
\arguments{
\item{mat}{a square matrix.}

\item{col}{the number of columns of the matrix.}
}
\value{
The sum of the matrix's diagonal and two off-diagonals within one cell from the diagonal.
}
\description{
Returns the sum of all cells on a grid's main diagonal (trace) and on its two off-diagonals.
}
