% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridProcessing.R
\name{classify_responses}
\alias{classify_responses}
\title{Form Counts for Four Grid Areas}
\usage{
classify_responses(grid, print_cell_class_mapping = FALSE)
}
\arguments{
\item{grid}{a matrix.}

\item{print_cell_class_mapping}{A logical indicating whether a matrix showing the mapping of response types to cells should be printed. If TRUE, then no computation occurs and any option passed to grid is ignored.}
}
\value{
A list containing the grid cell counts for each of four categories:
  indifferent, ambivalent, positive, and negative.
}
\description{
Returns the total cell counts for each of four grid regions corresponding to 
attitudes that can be measured using grid items.
}
