% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridProcessing.R
\name{create_grid_score}
\alias{create_grid_score}
\title{Form Likert-type Grid Scores}
\usage{
create_grid_score(
  x,
  gridinfo,
  b = (uni.min - uni.max)/16,
  uni.min = 1,
  uni.max = 9,
  reverse_code = NULL,
  subset = NULL,
  prefix = "conv_uni_"
)
}
\arguments{
\item{x}{data frame of grid-only LimeSurvey formatted data.}

\item{gridinfo}{list. The output of \code{grid_item_info}.}

\item{b}{a parameter between -1 and 0, default is a value satisfying Audrezet's (2016) six constraints.}

\item{uni.min, uni.max}{integers specifying the minimum and maximum value for the unidimensional (Likert-type) mapping. Default is a 9-point scale.}

\item{reverse_code}{a vector indicating which grids are reverse-coded.}

\item{subset}{a vector indiciating which items to include. By default (NULL) items are used.}

\item{prefix}{a string which is the prefix used for column names for the converted score data frame. 
a value of 0 means not reverse coded; a value of 1 means reverse coded.}
}
\value{
A data frame in which the grid scores for each respondent have been 
  converted to values equivalent to those on Likert-type scale.
}
\description{
For each item and respondent, creates scores equivalent to a Likert-type scale
out of LimeSurvey formatted grid item data.
}
