% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridProcessing.R
\name{create_grid_score}
\alias{create_grid_score}
\title{Form Likert-type Grid Scores}
\usage{
create_grid_score(x, gridinfo, b = -0.5, reverse_code = NULL)
}
\arguments{
\item{x}{data frame of grid-only LimeSurvey formatted data.}

\item{gridinfo}{list. The output of \code{grid_item_info}.}

\item{b}{a parameter between -1 and 0.}

\item{reverse_code}{a vector indicating which grids are reverse-coded. 
a value of 0 means not reverse coded; a value of 1 means reverse coded.}
}
\value{
A data frame in which the grid scores for each respondent have been 
  converted to values equivalent to those on 9 point Likert-type scales.
}
\description{
For each item and respondent, creates scores equivalent to a Likert-type scale
out of LimeSurvey formatted grid item data.
}
