% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridProcessing.R
\name{fix_limesurvey_likert}
\alias{fix_limesurvey_likert}
\title{Format LimeSurvey Likert-type Data}
\usage{
fix_limesurvey_likert(
  dat,
  mapping = "DA9",
  lsvals = NULL,
  livals = NULL,
  cols = NULL
)
}
\arguments{
\item{dat}{data frame of LimeSurvey formatted Likert-type data.}

\item{mapping}{character indicating which pre-generated mapping from LimeSurvey notation to numeric values should be used.}

\item{lsvals}{character vector of LimeSurvey's notation for Likert-type scale points.}

\item{livals}{integer vector of Likert-type scale points.}

\item{cols}{integer vector indicating which columns/items the fix should be applied to. If default (NULL), all columns/items will be used.}
}
\value{
A data frame of numeric Likert-type values for each response.
}
\description{
Replaces LimeSurvey's formatting of Likert-type data with the standard numerical scale points.
}
