% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridProcessing.R
\name{fix_limesurvey_likert}
\alias{fix_limesurvey_likert}
\title{Title}
\usage{
fix_limesurvey_likert(
  dat,
  lsvals = c("D4", "D3", "D2", "D1", "N0", "A1", "A2", "A3", "A4"),
  livals = 1:9
)
}
\arguments{
\item{dat}{}

\item{lsvals}{}

\item{livals}{}
}
\value{

}
\description{
Title
}
