% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrinomialTest.R
\name{generate_trinomial_info}
\alias{generate_trinomial_info}
\title{Title}
\usage{
generate_trinomial_info(
  n,
  alpha = 0.05,
  include_one = TRUE,
  find_RR = TRUE,
  digits = NULL,
  P0s = NULL
)
}
\arguments{
\item{n}{number of paired observations.}

\item{alpha}{significance level.}

\item{include_one}{logical. If \code{FALSE} the probability that all pairs are tied is excluded from \code{P0s}.}

\item{find_RR}{logical. If \code{TRUE} the output will include the rejection region.}

\item{digits}{number of significant digits to be returned.}

\item{P0s}{numeric vector of all possible probabilities of a tie.}
}
\value{

}
\description{
Title
}
