% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridProcessing.R
\name{grid_cell_counts}
\alias{grid_cell_counts}
\title{Title}
\usage{
grid_cell_counts(
  x,
  gridinfo,
  type = "items",
  reverse_code = NULL,
  return.table = FALSE,
  chosen_items = NULL
)
}
\arguments{
\item{x}{}

\item{gridinfo}{}

\item{type}{}

\item{reverse_code}{}

\item{return.table}{}

\item{chosen_items}{}
}
\value{

}
\description{
Title
}
