% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridProcessing.R
\name{grid_cell_counts}
\alias{grid_cell_counts}
\title{Title}
\usage{
grid_cell_counts(
  x,
  gridinfo,
  type = "items",
  reverse_code = NULL,
  return_table = FALSE,
  chosen_items = NULL
)
}
\arguments{
\item{x}{data frame of grid-only LimeSurvey formatted data.}

\item{gridinfo}{list. The output of \code{grid_item_info}.}

\item{type}{character. Indicates whether to create cell counts for \code{"items"} or \code{"respondents"}.}

\item{reverse_code}{a value of 1 indicates that the grid items are reverse coded, and will correct for it.}

\item{return_table}{logical. If \code{TRUE} a list of tables rather than matrices is returned.}

\item{chosen_items}{integer vector indicating which grid items to include.}
}
\value{
A list of matrices, each of which is either the sum of all responses to a certain item (the default)
  or the sum of a certain respondent's answers for all items.
}
\description{
Title
}
