% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridProcessing.R
\name{grid_item_info}
\alias{grid_item_info}
\title{Title}
\usage{
grid_item_info(x, rows = 5, cols = 5)
}
\arguments{
\item{x}{data frame of raw LimeSurvey formatted data.}

\item{rows, cols}{the dimensions of the grid items.}
}
\value{

}
\description{
Title
}
