% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridProcessing.R
\name{grid_summary_tri}
\alias{grid_summary_tri}
\title{Title}
\usage{
grid_summary_tri(
  mat,
  rows = 5,
  cols = 5,
  offdiag = 0,
  return.table = TRUE,
  limesurvey = TRUE
)
}
\arguments{
\item{mat}{}

\item{rows}{}

\item{cols}{}

\item{offdiag}{}

\item{return.table}{}

\item{limesurvey}{}
}
\value{

}
\description{
Title
}
