% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridProcessing.R
\name{grid_to_nine}
\alias{grid_to_nine}
\title{Title}
\usage{
grid_to_nine(gc, rows = 5, cols = 5, b = -0.5, rc = FALSE)
}
\arguments{
\item{gc}{}

\item{rows}{}

\item{cols}{}

\item{b}{}

\item{rc}{}
}
\value{

}
\description{
Title
}
