% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridProcessing.R
\name{grid_to_nine}
\alias{grid_to_nine}
\title{Title}
\usage{
grid_to_nine(gc, rows = 5, cols = 5, b = -0.5, rc = FALSE)
}
\arguments{
\item{gc}{integer indicating a grid cell as formatted by LimeSurvey.}

\item{rows, cols}{the dimensions of the grid item.}

\item{b}{a parameter between -1 and 0.}

\item{rc}{logical. If \code{TRUE} the function will correct for a reverse-coded grid item.}
}
\value{

}
\description{
Title
}
