% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphingFunctions.R
\name{make_heatmap}
\alias{make_heatmap}
\title{Make a Heatmap for a Grid}
\usage{
make_heatmap(
  grid,
  title,
  labels = c("agree_disagree", "satisfied_dissatisfied", "negative_positive")
)
}
\arguments{
\item{grid}{a matrix.}

\item{title}{the title of the plot.}

\item{labels}{a character string specifying which set of rownames and colnames
are to be used.}
}
\value{
a heatmap showing counts in each grid cell.
}
\description{
Make a Heatmap for a Grid
}
