% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphingFunctions.R
\name{make_path_diagram}
\alias{make_path_diagram}
\title{Make a Path Diagram for Two Grids}
\usage{
make_path_diagram(
  grid_data,
  grid_info,
  chosen_item_1,
  chosen_item_2,
  labels = c("agree_disagree", "satisfied_dissatisfied", "positive_negative", "other"),
  pos_labels = NULL,
  neg_labels = NULL,
  custom_x_name = NULL,
  custom_y_name = NULL
)
}
\arguments{
\item{grid_data}{data frame of grid-only LimeSurvey formatted data.}

\item{grid_info}{list. The output of \code{grid_item_info}.}

\item{chosen_item_1}{number. The first item of interest.}

\item{chosen_item_2}{number. The second item of interest.}

\item{labels}{a character string specifying which set of rownames and colnames
are to be used.}

\item{pos_labels}{character. Custom column names.}

\item{neg_labels}{character. Custom row names.}

\item{custom_x_name}{Custom name for the x-axis.}

\item{custom_y_name}{Custom name for the y-axis.}
}
\value{
a path diagram showing the shift in each respondent's answer from one
  item to the other.
}
\description{
Make a Path Diagram for Two Grids
}
