% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphingFunctions.R
\name{make_path_diagram}
\alias{make_path_diagram}
\title{Title}
\usage{
make_path_diagram(
  grid_data,
  grid_info,
  chosen_item_1,
  chosen_item_2,
  labels = c("agree_disagree", "satisfied_dissatisfied", "negative_positive")
)
}
\arguments{
\item{grid_data}{data frame of grid-only LimeSurvey formatted data.}

\item{grid_info}{list. The output of \code{grid_item_info}.}

\item{chosen_item_1}{number. The first item of interest.}

\item{chosen_item_2}{number. The second item of interest.}

\item{labels}{a character string specifying which set of rownames and colnames
are to be used.}
}
\value{
a path diagram showing the shift in each respondent's answer from one
  item to the other.
}
\description{
Title
}
