% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrinomialTest.R
\name{trinomial.test}
\alias{trinomial.test}
\title{Title}
\usage{
trinomial.test(
  col1,
  col2 = NULL,
  alternative = c("two.sided", "greater", "less"),
  p_tie = NULL,
  print.info = TRUE
)
}
\arguments{
\item{col1}{first data vector or a vector of (positive, tie, negative) counts of differences}

\item{col2}{second data vector}

\item{alternative}{the type of hypothesis test to be performed}

\item{p_tie}{theoretical proportion of ties; calculated from the data by default}

\item{print.info}{}
}
\value{

}
\description{
Title
}
