% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridProcessing.R
\name{within_diag}
\alias{within_diag}
\title{Title}
\usage{
within_diag(mat, col = 5)
}
\arguments{
\item{mat}{a square matrix.}

\item{col}{the number of columns of the matrix.}
}
\value{

}
\description{
Title
}
